# Conflicts:
#	src/Model/Table/SerialNumbersTable.php

+++ src/Model/Table/SalesOrderRowsTable.php
@@ -81,7 +81,7 @@ class SalesOrderRowsTable extends Table
 
 
         $validator
-            ->integer('quantity')
+            ->decimal('quantity')
             ->requirePresence('quantity', 'create')
             ->notEmpty('quantity');
